refactored for modularity
